Практика по заданию 5.57

+++ Task_5.5.5/Program.cs
@@ -7,7 +7,7 @@
             Console.WriteLine(Factorial(20));
         }
 
-        static int Factorial(int x)
+        static decimal Factorial(int x)
         {
             if (x == 0)
             {
